refactor and improve repo type handling
