+++ repos/utils.py
@@ -1,5 +1,5 @@
 # Copyright 2012 VPAC, http://www.vpac.org
-# Copyright 2013-2021 Marcus Furlong <furlongm@gmail.com>
+# Copyright 2013-2025 Marcus Furlong <furlongm@gmail.com>
 #
 # This file is part of Patchman.
 #
@@ -15,37 +15,24 @@
 # You should have received a copy of the GNU General Public License
 # along with Patchman. If not, see <http://www.gnu.org/licenses/>
 
-import io
-import os
-import git
 import re
-import shutil
-import tarfile
-import tempfile
-import yaml
 from io import BytesIO
 
 from defusedxml import ElementTree as ET
-from debian.debian_support import Version
-from debian.deb822 import Packages
-from fnmatch import fnmatch
 from tenacity import RetryError
 
 from django.db import IntegrityError, DatabaseError, transaction
 from django.db.models import Q
 
-from arch.models import PackageArchitecture
-from packages.models import Package, PackageString
-from packages.utils import parse_package_string, get_or_create_package, find_evr, \
-    convert_package_to_packagestring, convert_packagestring_to_package
-from util import get_url, download_url, response_is_valid, extract, get_checksum, Checksum, get_setting_of_type, \
-    get_datetime_now
+from packages.models import Package
+from packages.utils import convert_package_to_packagestring, convert_packagestring_to_package
+from util import get_url, download_url, response_is_valid, extract, get_checksum, Checksum, get_setting_of_type
 from patchman.signals import info_message, warning_message, error_message, debug_message, pbar_start, pbar_update
 
 
 def get_or_create_repo(r_name, r_arch, r_type, r_id=None):
-    """ Get or create a Repository object. Returns the object. Returns None if
-        it cannot get or create the object.
+    """ Get or create a Repository object and returns the object.
+        Returns None if it cannot get or create the object.
     """
     from repos.models import Repository
     repositories = Repository.objects.all()
@@ -105,33 +92,6 @@ def update_mirror_packages(mirror, packages):
     mirror.save()
 
 
-def get_repomd_url(mirror_url, data, url_type='primary'):
-    if isinstance(data, str):
-        if data.startswith('Bad repo - not in list') or data.startswith('Invalid repo'):
-            return None, None, None
-
-    ns = 'http://linux.duke.edu/metadata/repo'
-    extracted = extract(data, mirror_url)
-    location = None
-    try:
-        tree = ET.parse(BytesIO(extracted))
-        root = tree.getroot()
-        for child in root:
-            if child.attrib.get('type') == url_type:
-                for grandchild in child:
-                    if grandchild.tag == f'{{{ns}}}location':
-                        location = grandchild.attrib.get('href')
-                    if grandchild.tag == f'{{{ns}}}checksum':
-                        checksum = grandchild.text
-                        checksum_type = grandchild.attrib.get('type')
-    except ET.ParseError as e:
-        error_message.send(sender=None, text=(f'Error parsing repomd from {mirror_url}: {e}'))
-    if not location:
-        return None, None, None
-    url = str(mirror_url.rsplit('/', 2)[0]) + '/' + location
-    return url, checksum, checksum_type
-
-
 def find_mirror_url(stored_mirror_url, formats):
     """ Find the actual URL of the mirror by trying predefined paths
     """
@@ -148,70 +108,6 @@ def find_mirror_url(stored_mirror_url, formats):
             continue
         if res is not None and res.ok:
             return res
-
-
-def get_gentoo_mirror_urls():
-    """ Use the Gentoo API to find http(s) mirrors
-    """
-    gentoo_distfiles_url = 'https://api.gentoo.org/mirrors/distfiles.xml'
-    res = get_url(gentoo_distfiles_url)
-    if not res:
-        return
-    mirrors = {}
-    try:
-        tree = ET.parse(BytesIO(res.content))
-        root = tree.getroot()
-        for child in root:
-            if child.tag == 'mirrorgroup':
-                for k, v in child.attrib.items():
-                    if k == 'region':
-                        region = v
-                    elif k == 'country':
-                        country = v
-                for mirror in child:
-                    for element in mirror:
-                        if element.tag == 'name':
-                            name = element.text
-                            mirrors[name] = {}
-                            mirrors[name]['region'] = region
-                            mirrors[name]['country'] = country
-                            mirrors[name]['urls'] = []
-                        elif element.tag == 'uri':
-                            if element.get('protocol') == 'http':
-                                mirrors[name]['urls'].append(element.text)
-    except ET.ParseError as e:
-        error_message.send(sender=None, text=f'Error parsing {gentoo_distfiles_url}: {e}')
-    mirror_urls = []
-    # for now, ignore region data and choose MAX_MIRRORS mirrors at random
-    for _, v in mirrors.items():
-        for url in v['urls']:
-            mirror_urls.append(url.rstrip('/') + '/snapshots/gentoo-latest.tar.xz')
-    return mirror_urls
-
-
-def get_gentoo_overlay_mirrors(repo_name):
-    """Get the gentoo overlay repos that match repo.id
-    """
-    gentoo_overlays_url = 'https://api.gentoo.org/overlays/repositories.xml'
-    res = get_url(gentoo_overlays_url)
-    if not res:
-        return
-    mirrors = []
-    try:
-        tree = ET.parse(BytesIO(res.content))
-        root = tree.getroot()
-        for child in root:
-            if child.tag == 'repo':
-                found = False
-                for element in child:
-                    if element.tag == 'name' and element.text == repo_name:
-                        found = True
-                    if found and element.tag == 'source':
-                        if element.text.startswith('http'):
-                            mirrors.append(element.text)
-    except ET.ParseError as e:
-        error_message.send(sender=None, text=f'Error parsing {gentoo_overlays_url}: {e}')
-    return mirrors
 
 
 def is_metalink(url):
@@ -268,7 +164,7 @@ def get_mirrorlist_urls(url):
                 return
             mirror_urls = re.findall(r'^http[s]*://.*$|^ftp://.*$', data.decode('utf-8'), re.MULTILINE)
             if mirror_urls:
-
+                debug_message.send(sender=None, text=f'Found mirrorlist: {url}')
                 return mirror_urls
             else:
                 debug_message.send(sender=None, text=f'Not a mirrorlist: {url}')
@@ -326,225 +222,8 @@ def check_for_metalinks(repo):
             mirror.mirrorlist = True
             mirror.last_access_ok = True
             mirror.save()
-            text = f'Found metalink - {mirror.url}'
-            info_message.send(sender=None, text=text)
+            info_message.send(sender=None, text=f'Found metalink - {mirror.url}')
             add_mirrors_from_urls(repo, mirror_urls)
-
-
-def extract_module_metadata(data, url, repo):
-    """ Extract module metadata from a modules.yaml file
-    """
-    modules = set()
-    extracted = extract(data, url)
-    try:
-        modules_yaml = yaml.safe_load_all(extracted)
-    except yaml.YAMLError as e:
-        error_message.send(sender=None, text=f'Error parsing modules.yaml: {e}')
-
-    mlen = len(re.findall(r'---', yaml.dump(extracted.decode())))
-    pbar_start.send(sender=None, ptext=f'Extracting {mlen} Modules ', plen=mlen)
-    for i, doc in enumerate(modules_yaml):
-        pbar_update.send(sender=None, index=i + 1)
-        document = doc['document']
-        modulemd = doc['data']
-        if document == 'modulemd':
-            modulemd = doc['data']
-            m_name = modulemd.get('name')
-            m_stream = modulemd['stream']
-            m_version = modulemd.get('version')
-            m_context = modulemd.get('context')
-            arch = modulemd.get('arch')
-            raw_packages = modulemd.get('artifacts', {}).get('rpms', '')
-            # raw_profiles = list(modulemd.get('profiles', {}).keys())
-
-            packages = set()
-            p_type = Package.RPM
-            for pkg_str in raw_packages:
-                p_name, p_epoch, p_ver, p_rel, p_dist, p_arch = parse_package_string(pkg_str)
-                package = get_or_create_package(p_name, p_epoch, p_ver, p_rel, p_arch, p_type)
-                packages.add(package)
-
-            from modules.utils import get_or_create_module
-            module, created = get_or_create_module(m_name, m_stream, m_version, m_context, arch, repo)
-
-            package_ids = []
-            for package in packages:
-                package_ids.append(package.id)
-                module.packages.add(package)
-            for package in module.packages.all():
-                if package.id not in package_ids:
-                    module.packages.remove(package)
-            modules.add(module)
-
-
-def extract_yum_packages(data, url):
-    """ Extract package metadata from a yum primary.xml file
-    """
-    extracted = extract(data, url)
-    ns = 'http://linux.duke.edu/metadata/common'
-    packages = set()
-    try:
-        context = ET.iterparse(BytesIO(extracted), events=('start', 'end'))
-        for event, elem in context:
-            if event == 'start':
-                if elem.tag == f'{{{ns}}}metadata':
-                    plen = int(elem.attrib.get('packages'))
-                    break
-        pbar_start.send(sender=None, ptext=f'Extracting {plen} Packages', plen=plen)
-        i = 0
-        for event, elem in context:
-            if event == 'start':
-                if elem.tag == f'{{{ns}}}package':
-                    if elem.attrib.get('type') == 'rpm':
-                        name = version = release = arch = ''
-            elif event == 'end':
-                if elem.tag == f'{{{ns}}}name':
-                    name = elem.text.lower()
-                elif elem.tag == f'{{{ns}}}arch':
-                    arch = elem.text
-                elif elem.tag == f'{{{ns}}}version':
-                    fullversion = elem
-                    epoch = fullversion.get('epoch')
-                    version = fullversion.get('ver')
-                    release = fullversion.get('rel')
-                elif elem.tag == f'{{{ns}}}package':
-                    if name and version and release and arch:
-                        if epoch == '0':
-                            epoch = ''
-                        package = PackageString(
-                            name=name,
-                            epoch=epoch,
-                            version=version,
-                            release=release,
-                            arch=arch,
-                            packagetype='R',
-                        )
-                        packages.add(package)
-                        pbar_update.send(sender=None, index=i + 1)
-                        i += 1
-                    else:
-                        text = f'Error parsing Package: {name} {epoch} {version} {release} {arch}'
-                        error_message.send(sender=None, text=text)
-                elem.clear()
-    except ET.ParseError as e:
-        error_message.send(sender=None, text=f'Error parsing yum primary.xml from {url}: {e}')
-    return packages
-
-
-def extract_deb_packages(data, url):
-    """ Extract package metadata from debian Packages file
-    """
-    try:
-        extracted = extract(data, url).decode('utf-8')
-    except UnicodeDecodeError as e:
-        error_message.send(sender=None, text=f'Skipping {url} : {e}')
-        return
-    package_re = re.compile('^Package: ', re.M)
-    plen = len(package_re.findall(extracted))
-    packages = set()
-
-    if plen > 0:
-        pbar_start.send(sender=None, ptext=f'Extracting {plen} Packages', plen=plen)
-        for i, stanza in enumerate(Packages.iter_paragraphs(extracted)):
-            # https://github.com/furlongm/patchman/issues/55
-            if 'version' not in stanza:
-                continue
-            fullversion = Version(stanza['version'])
-            arch = stanza['architecture']
-            name = stanza['package']
-            epoch = fullversion._BaseVersion__epoch
-            if epoch is None:
-                epoch = ''
-            version = fullversion._BaseVersion__upstream_version
-            release = fullversion._BaseVersion__debian_revision
-            if release is None:
-                release = ''
-            pbar_update.send(sender=None, index=i + 1)
-            package = PackageString(name=name,
-                                    epoch=epoch,
-                                    version=version,
-                                    release=release,
-                                    arch=arch,
-                                    packagetype='D')
-            packages.add(package)
-    else:
-        info_message.send(sender=None, text='No packages found in repo')
-    return packages
-
-
-def extract_yast_packages(data):
-    """ Extract package metadata from yast metadata file
-    """
-    extracted = extract(data, 'gz').decode('utf-8')
-    pkgs = re.findall('=Pkg: (.*)', extracted)
-    plen = len(pkgs)
-    packages = set()
-
-    if plen > 0:
-        pbar_start.send(sender=None, ptext=f'Extracting {plen} Packages', plen=plen)
-
-        for i, pkg in enumerate(pkgs):
-            pbar_update.send(sender=None, index=i + 1)
-            name, version, release, arch = pkg.split()
-            package = PackageString(name=name.lower(),
-                                    epoch='',
-                                    version=version,
-                                    release=release,
-                                    arch=arch,
-                                    packagetype='R')
-            packages.add(package)
-    else:
-        info_message.send(sender=None, text='No packages found in repo')
-    return packages
-
-
-def extract_arch_packages(data):
-    """ Extract package metadata from an arch linux tarfile
-    """
-    from packages.utils import find_evr
-    bio = BytesIO(data)
-    tf = tarfile.open(fileobj=bio, mode='r:*')
-    packages = set()
-    plen = len(tf.getnames())
-    if plen > 0:
-        pbar_start.send(sender=None, ptext=f'Extracting {plen} Packages', plen=plen)
-        for i, tarinfo in enumerate(tf):
-            pbar_update.send(sender=None, index=i + 1)
-            if tarinfo.isfile():
-                name_sec = ver_sec = arch_sec = False
-                t = tf.extractfile(tarinfo).read()
-                for line in t.decode('utf-8').splitlines():
-                    if line.startswith('%NAME%'):
-                        name_sec = True
-                        continue
-                    if name_sec:
-                        name_sec = False
-                        name = line
-                        continue
-                    if line.startswith('%VERSION%'):
-                        ver_sec = True
-                        continue
-                    if ver_sec:
-                        ver_sec = False
-                        epoch, version, release = find_evr(line)
-                        continue
-                    if line.startswith('%ARCH%'):
-                        arch_sec = True
-                        continue
-                    if arch_sec:
-                        arch_sec = False
-                        arch = line
-                        continue
-                package = PackageString(name=name.lower(),
-                                        epoch=epoch,
-                                        version=version,
-                                        release=release,
-                                        arch=arch,
-                                        packagetype='A')
-                packages.add(package)
-    else:
-        info_message.send(sender=None, text='No packages found in repo')
-    return packages
 
 
 def fetch_mirror_data(mirror, url, text, checksum=None, checksum_type=None, metadata_type=None):
@@ -591,515 +270,6 @@ def mirror_checksum_is_valid(computed, provided, mirror, metadata_type):
         return False
     else:
         return True
-
-
-def extract_updateinfo(data, url):
-    """ Parses updateinfo.xml and extracts package/errata information
-    """
-    print(url)
-    from errata.utils import get_or_create_erratum
-    extracted = extract(data, url)
-    updates = []
-    try:
-        tree = ET.parse(BytesIO(extracted))
-        root = tree.getroot()
-        elen = root.__len__()
-        pbar_start.send(sender=None, ptext=f'Extracting {elen} rpm Errata', plen=elen)
-        for i, update in enumerate(root.findall('update')):
-            pbar_update.send(sender=None, index=i + 1)
-            e_type = update.attrib.get('type')
-            name = update.find('id').text
-            synopsis = update.find('title').text
-            issue_date = update.find('issued').attrib.get('date')
-            e, created = get_or_create_erratum(name, e_type, issue_date, synopsis)
-
-            xreferences = update.find('references')
-            for reference in xreferences.findall('reference'):
-                if reference.attrib.get('type') == 'cve':
-                    cve_id = reference.attrib.get('id')
-                    e.add_cve(cve_id)
-                else:
-                    ref = reference.attrib.get('href')
-                    e.add_reference('Link', ref)
-
-            osrelease_name = None
-            release = update.find('release')
-            if release:
-                osrelease_name = release.text
-
-            pkglist = update.find('pkglist')
-            packages = set()
-            for collection in pkglist.findall('collection'):
-                if not osrelease_name:
-                    collection_name = collection.find('name')
-                    if collection_name is not None:
-                        osrelease_name = collection_name.text
-                from operatingsystems.models import OSRelease
-                if osrelease_name:
-                    osrelease, created = OSRelease.objects.get_or_create(name=osrelease_name)
-                    e.osreleases.add(osrelease)
-                    # TODO for opensuse, add if repo is associated with an os release
-
-                for pkg in collection.findall('package'):
-                    name = pkg.attrib.get('name')
-                    epoch = pkg.attrib.get('epoch')
-                    version = pkg.attrib.get('version')
-                    release = pkg.attrib.get('release')
-                    arch = pkg.attrib.get('arch')
-                    package = get_or_create_package(
-                        name=name.lower(),
-                        epoch=epoch,
-                        version=version,
-                        release=release,
-                        arch=arch,
-                        p_type='R',
-                    )
-                    packages.add(package)
-                e.add_packages(packages)
-                e.save()
-    except ET.ParseError as e:
-        error_message.send(sender=None, text=f'Error parsing updateinfo file: {e}')
-    return updates
-
-
-def refresh_rpm_updateinfo(mirror, data, mirror_url):
-    url, checksum, checksum_type = get_repomd_url(mirror_url, data, url_type='updateinfo')
-    data = fetch_mirror_data(
-        mirror=mirror,
-        url=url,
-        checksum=checksum,
-        checksum_type=checksum_type,
-        text='Downloading Errata data',
-        metadata_type='updateinfo')
-
-    if not mirror.last_access_ok:
-        return
-
-    if mirror.modules_checksum == checksum:
-        text = 'Mirror Errata checksum has not changed, skipping Erratum refresh'
-        warning_message.send(sender=None, text=text)
-        return
-    else:
-        mirror.modules_checksum = checksum
-        mirror.save()
-
-    extract_updateinfo(data, url)
-
-
-def refresh_rpm_modules(mirror, data, mirror_url):
-    url, checksum, checksum_type = get_repomd_url(mirror_url, data, url_type='modules')
-    if url:
-        data = fetch_mirror_data(
-            mirror=mirror,
-            url=url,
-            checksum=checksum,
-            checksum_type=checksum_type,
-            text='Downloading Module data',
-            metadata_type='module')
-
-    if not mirror.last_access_ok:
-        return
-
-    if mirror.modules_checksum == checksum:
-        text = 'Mirror Modules checksum has not changed, skipping Module refresh'
-        warning_message.send(sender=None, text=text)
-        return
-    else:
-        mirror.modules_checksum = checksum
-        mirror.save()
-
-    extract_module_metadata(data, url, mirror.repo)
-
-
-def refresh_rpm_primary(mirror, data, mirror_url, ts):
-    url, checksum, checksum_type = get_repomd_url(mirror_url, data, url_type='primary')
-    data = fetch_mirror_data(
-        mirror=mirror,
-        url=url,
-        checksum=checksum,
-        checksum_type=checksum_type,
-        text='Downloading Package data',
-        metadata_type='package')
-
-    if not mirror.last_access_ok:
-        return
-
-    if mirror.packages_checksum == checksum:
-        text = 'Mirror Packages checksum has not changed, skipping Package refresh'
-        warning_message.send(sender=None, text=text)
-        return
-    else:
-        mirror.packages_checksum = checksum
-        mirror.save()
-
-    # only refresh X mirrors, where X = max_mirrors
-    max_mirrors = get_max_mirrors()
-    mirrors_q = Q(mirrorlist=False, refresh=True, enabled=True, timestamp=ts, packages_checksum=checksum)
-    have_checksum_and_ts = mirror.repo.mirror_set.filter(mirrors_q).count()
-    if have_checksum_and_ts >= max_mirrors:
-        text = f'{max_mirrors} Mirrors already have this checksum and timestamp, skipping Package refresh'
-        info_message.send(sender=None, text=text)
-        return
-
-    packages = extract_yum_packages(data, url)
-    if packages:
-        update_mirror_packages(mirror, packages)
-
-
-def refresh_yum_repo(mirror, data, mirror_url, ts, errata_only):
-    """ Refresh package, module and updateinfo/errata data for a yum-style rpm Mirror
-    """
-    if not errata_only:
-        refresh_rpm_primary(mirror, data, mirror_url, ts)
-        refresh_rpm_modules(mirror, data, mirror_url)
-    refresh_rpm_updateinfo(mirror, data, mirror_url)
-
-
-def refresh_yum_repo_errata(repo):
-    refresh_rpm_repo_mirrors(repo, errata_only=True)
-
-
-def refresh_arch_repo(repo):
-    """ Refresh all mirrors of an arch linux repo
-    """
-    max_mirrors = get_max_mirrors()
-    fname = f'{repo.arch}/{repo.repo_id}.db'
-    ts = get_datetime_now()
-
-    enabled_mirrors = repo.mirror_set.filter(refresh=True, enabled=True)
-    for i, mirror in enumerate(enabled_mirrors):
-        if i >= max_mirrors:
-            text = f'{max_mirrors} Mirrors already refreshed (max={max_mirrors}), skipping further refreshes'
-            warning_message.send(sender=None, text=text)
-            break
-
-        res = find_mirror_url(mirror.url, [fname])
-        if not res:
-            continue
-        mirror_url = res.url
-        text = f'Found Arch Repo - {mirror_url}'
-        info_message.send(sender=None, text=text)
-
-        package_data = fetch_mirror_data(
-            mirror=mirror,
-            url=mirror_url,
-            text='Downloading Repo data')
-        if not package_data:
-            continue
-
-        computed_checksum = get_checksum(package_data, Checksum.sha1)
-        if mirror.packages_checksum == computed_checksum:
-            text = 'Mirror checksum has not changed, not refreshing Package metadata'
-            warning_message.send(sender=None, text=text)
-            continue
-        else:
-            mirror.packages_checksum = computed_checksum
-
-        packages = extract_arch_packages(package_data)
-        update_mirror_packages(mirror, packages)
-        packages.clear()
-        mirror.timestamp = ts
-        mirror.save()
-
-
-def refresh_gentoo_main_repo(repo):
-    """ Refresh all mirrors of the main gentoo repo
-    """
-    mirrors = get_gentoo_mirror_urls()
-    add_mirrors_from_urls(repo, mirrors)
-
-
-def refresh_gentoo_overlay_repo(repo):
-    """ Refresh all mirrors of a Gentoo overlay repo
-    """
-    mirrors = get_gentoo_overlay_mirrors(repo.repo_id)
-    add_mirrors_from_urls(repo, mirrors)
-
-
-def get_gentoo_ebuild_keywords(content):
-    keywords = set()
-    default_keywords = {
-        'alpha',
-        'amd64',
-        'arm',
-        'arm64',
-        'hppa',
-        'loong',
-        'm68k',
-        'mips',
-        'ppc',
-        'ppc64',
-        'riscv',
-        's390',
-        'sparc',
-        'x86',
-    }
-    for line in content.decode().splitlines():
-        if not line.startswith('KEYWORDS='):
-            continue
-        all_keywords = line.split('=')[1].split('#')[0].strip(' "').split()
-        if len(all_keywords) == 0 or '*' in all_keywords:
-            all_keywords = default_keywords
-        for keyword in all_keywords:
-            if keyword.startswith('~'):
-                continue
-            if keyword.startswith('-'):
-                keyword = keyword.replace('-', '')
-                if keyword in all_keywords:
-                    all_keywords.remove(keyword)
-                continue
-            keywords.add(keyword)
-        break
-    return keywords
-
-
-def extract_gentoo_ebuilds(data):
-    extracted_ebuilds = {}
-    with tarfile.open(fileobj=io.BytesIO(data), mode='r') as tar:
-        for member in tar.getmembers():
-            if member.isfile() and member.name.endswith('ebuild') and not member.name.endswith('skel.ebuild'):
-                file_content = tar.extractfile(member).read()
-                extracted_ebuilds[member.name] = file_content
-    return extracted_ebuilds
-
-
-def extract_gentoo_packages(mirror, data):
-    extracted_ebuilds = extract_gentoo_ebuilds(data)
-    return extract_gentoo_packages_from_ebuilds(extracted_ebuilds)
-
-
-def extract_gentoo_packages_from_ebuilds(extracted_ebuilds):
-    if not extracted_ebuilds:
-        return
-
-    packages = set()
-    flen = len(extracted_ebuilds)
-    pbar_start.send(sender=None, ptext=f'Processing {flen} ebuilds', plen=flen)
-    for i, (path, content) in enumerate(extracted_ebuilds.items()):
-        pbar_update.send(sender=None, index=i + 1)
-        components = path.split(os.sep)
-        category = components[1]
-        name = components[2]
-        evr = components[3].replace(f'{name}-', '').replace('.ebuild', '')
-        epoch, version, release = find_evr(evr)
-        arches = get_gentoo_ebuild_keywords(content)
-        for arch in arches:
-            package = PackageString(
-                name=name.lower(),
-                epoch=epoch,
-                version=version,
-                release=release,
-                arch=arch,
-                packagetype='G',
-                category=category,
-            )
-            packages.add(package)
-    plen = len(packages)
-    info_message.send(sender=None, text=f'Extracted {plen} Packages', plen=plen)
-    return packages
-
-
-def extract_gentoo_overlay_packages(mirror):
-    t = tempfile.mkdtemp()
-    git.Repo.clone_from(mirror.url, t, branch='master', depth=1)
-    packages = set()
-    arch, c = PackageArchitecture.objects.get_or_create(name='any')
-    for root, dirs, files in os.walk(t):
-        for name in files:
-            if fnmatch(name, '*.ebuild'):
-                full_name = root.replace(t + '/', '')
-                p_category, p_name = full_name.split('/')
-                m = re.match(fr'{p_name}-(.*)\.ebuild', name)
-                if m:
-                    p_evr = m.group(1)
-                epoch, version, release = find_evr(p_evr)
-                package = PackageString(
-                    name=p_name.lower(),
-                    epoch=epoch,
-                    version=version,
-                    release=release,
-                    arch=arch,
-                    packagetype='G',
-                    category=p_category,
-                )
-                packages.add(package)
-    shutil.rmtree(t)
-    return packages
-
-
-def refresh_gentoo_repo(repo):
-    """ Refresh a Gentoo repo
-    """
-    if repo.repo_id == 'gentoo':
-        repo_type = 'main'
-        refresh_gentoo_main_repo(repo)
-    else:
-        refresh_gentoo_overlay_repo(repo)
-        repo_type = 'overlay'
-    ts = get_datetime_now()
-    for mirror in repo.mirror_set.filter(mirrorlist=False, refresh=True, enabled=True):
-        res = get_url(mirror.url + '.md5sum')
-        data = download_url(res, 'Downloading Repo checksum')
-        if data is None:
-            mirror.fail()
-            continue
-        checksum = data.decode().split()[0]
-        if checksum is None:
-            mirror.fail()
-            continue
-        if mirror.packages_checksum == checksum:
-            text = 'Mirror checksum has not changed, not refreshing Package metadata'
-            warning_message.send(sender=None, text=text)
-            continue
-        res = get_url(mirror.url)
-        mirror.last_access_ok = response_is_valid(res)
-        if mirror.last_access_ok:
-            data = download_url(res, 'Downloading Repo data')
-            if data is None:
-                mirror.fail()
-                continue
-            extracted = extract(data, mirror.url)
-            text = f'Found Gentoo Repo - {mirror.url}'
-            info_message.send(sender=None, text=text)
-            computed_checksum = get_checksum(data, Checksum.md5)
-            if not mirror_checksum_is_valid(computed_checksum, checksum, mirror, 'package'):
-                continue
-            else:
-                mirror.packages_checksum = checksum
-            if repo_type == 'main':
-                packages = extract_gentoo_packages(mirror, extracted)
-            elif repo_type == 'overlay':
-                packages = extract_gentoo_overlay_packages(mirror)
-            mirror.timestamp = ts
-            if packages:
-                update_mirror_packages(mirror, packages)
-        else:
-            mirror.fail()
-        mirror.save()
-
-
-def refresh_yast_repo(mirror, data):
-    """ Refresh package metadata for a yast-style rpm mirror
-        and add the packages to the mirror
-    """
-    package_dir = re.findall('DESCRDIR *(.*)', data.decode('utf-8'))[0]
-    package_url = f'{mirror.url}/{package_dir}/packages.gz'
-
-    package_data = fetch_mirror_data(
-        mirror=mirror,
-        url=package_url,
-        text='Downloading yast Repo data')
-    if not package_data:
-        return
-
-    mirror.packages_checksum = 'yast'
-    packages = extract_yast_packages(package_data)
-    if packages:
-        update_mirror_packages(mirror, packages)
-        packages.clear()
-
-
-def refresh_rpm_repo(repo):
-    """ Refresh an rpm repo (yum or yast)
-        Checks if the repo url is a mirrorlist or metalink,
-        and extracts mirrors if so, then refreshes the mirrors
-    """
-    check_for_mirrorlists(repo)
-    check_for_metalinks(repo)
-    refresh_rpm_repo_mirrors(repo)
-
-
-def refresh_rpm_repo_mirrors(repo, errata_only=False):
-    """ checks a number of common yum repo formats to determine
-        which type of repo it is, then refreshes the mirrors
-    """
-    formats = [
-        'repodata/repomd.xml.xz',
-        'repodata/repomd.xml.bz2',
-        'repodata/repomd.xml.gz',
-        'repodata/repomd.xml',
-        'suse/repodata/repomd.xml.xz',
-        'suse/repodata/repomd.xml.bz2',
-        'suse/repodata/repomd.xml.gz',
-        'suse/repodata/repomd.xml',
-        'content',
-    ]
-    max_mirrors = get_max_mirrors()
-    ts = get_datetime_now()
-    enabled_mirrors = repo.mirror_set.filter(mirrorlist=False, refresh=True, enabled=True)
-    for i, mirror in enumerate(enabled_mirrors):
-        if i >= max_mirrors:
-            text = f'{max_mirrors} Mirrors already refreshed (max={max_mirrors}), skipping further refreshes'
-            warning_message.send(sender=None, text=text)
-            break
-
-        res = find_mirror_url(mirror.url, formats)
-        if not res:
-            mirror.fail()
-            continue
-        mirror_url = res.url
-
-        repo_data = fetch_mirror_data(
-            mirror=mirror,
-            url=mirror_url,
-            text='Downloading Repo data')
-        if not repo_data:
-            continue
-
-        if mirror_url.endswith('content'):
-            text = f'Found yast rpm Repo - {mirror_url}'
-            info_message.send(sender=None, text=text)
-            refresh_yast_repo(mirror, repo_data)
-        else:
-            text = f'Found yum rpm Repo - {mirror_url}'
-            info_message.send(sender=None, text=text)
-            refresh_yum_repo(mirror, repo_data, mirror_url, ts, errata_only)
-        mirror.timestamp = ts
-        mirror.save()
-
-
-def refresh_deb_repo(repo):
-    """ Refresh a debian repo.
-        Checks for the Packages* files to determine what the mirror urls
-        are and then downloads and extracts packages from those files.
-    """
-
-    formats = ['Packages.xz', 'Packages.bz2', 'Packages.gz', 'Packages']
-
-    ts = get_datetime_now()
-    enabled_mirrors = repo.mirror_set.filter(refresh=True, enabled=True)
-    for mirror in enabled_mirrors:
-        res = find_mirror_url(mirror.url, formats)
-        if not res:
-            continue
-        mirror_url = res.url
-        text = f'Found deb Repo - {mirror_url}'
-        info_message.send(sender=None, text=text)
-
-        package_data = fetch_mirror_data(
-            mirror=mirror,
-            url=mirror_url,
-            text='Downloading Repo data')
-        if not package_data:
-            continue
-
-        computed_checksum = get_checksum(package_data, Checksum.sha1)
-        if mirror.packages_checksum == computed_checksum:
-            text = 'Mirror checksum has not changed, not refreshing Package metadata'
-            warning_message.send(sender=None, text=text)
-            continue
-        else:
-            mirror.packages_checksum = computed_checksum
-
-        packages = extract_deb_packages(package_data, mirror_url)
-        if not packages:
-            mirror.fail()
-            continue
-
-        update_mirror_packages(mirror, packages)
-        packages.clear()
-        mirror.timestamp = ts
-        mirror.save()
 
 
 def find_best_repo(package, hostrepos):
